separate ner and pos
