+++ notebooks/token_classification_notes.md
@@ -1,0 +1,84 @@
+thainer pos_pags best 1.0 0.0
+              precision    recall  f1-score   support
+
+         ADJ     0.9027    0.7351    0.8104       404
+         ADP     0.9584    0.9294    0.9437      1190
+         ADV     0.8839    0.6859    0.7724       433
+         AUX     0.9592    0.9697    0.9644      1090
+       CCONJ     0.9810    0.9663    0.9736       534
+         DET     0.9471    0.8564    0.8994       376
+        NOUN     0.9189    0.9747    0.9460      9324
+         NUM     0.9783    0.9082    0.9419       795
+        PART     0.9897    0.9570    0.9731       302
+        PRON     0.9627    0.9281    0.9451       167
+       PROPN     0.9157    0.7937    0.8503       698
+       PUNCT     0.9980    0.9980    0.9980      4423
+       SCONJ     0.9655    0.9444    0.9549      1422
+        VERB     0.9326    0.8993    0.9156      3603
+
+    accuracy                         0.9450     24761
+   macro avg     0.9496    0.8962    0.9206     24761
+weighted avg     0.9452    0.9450    0.9442     24761
+
+thainer ner_tags best 1.0 0.0
+              precision    recall  f1-score   support
+
+        DATE     0.8941    0.8216    0.8563       185
+       EMAIL     1.0000    1.0000    1.0000         1
+         LAW     0.8000    0.5333    0.6400        15
+         LEN     0.9444    0.8095    0.8718        21
+    LOCATION     0.7722    0.6667    0.7156       483
+       MONEY     0.9808    0.8644    0.9189        59
+ORGANIZATION     0.8714    0.7287    0.7937       586
+     PERCENT     1.0000    0.8667    0.9286        15
+      PERSON     0.8876    0.7816    0.8312       293
+       PHONE     0.7500    0.6000    0.6667        10
+        TIME     0.8333    0.6383    0.7229        94
+         URL     1.0000    0.8571    0.9231         7
+         ZIP     1.0000    0.5000    0.6667         2
+
+   micro avg     0.8529    0.7301    0.7867      1771
+   macro avg     0.9026    0.7437    0.8104      1771
+weighted avg     0.8524    0.7301    0.7861      1771
+
+lst20 pos_tags best 1.0 0.0
+              precision    recall  f1-score   support
+
+          NN     0.9699    0.9780    0.9740     58568
+          VV     0.9567    0.9670    0.9618     42586
+          PU     0.9999    0.9998    0.9999     37973
+          CC     0.9485    0.9671    0.9577     17613
+          PS     0.9413    0.9421    0.9417     10886
+          AX     0.9514    0.9427    0.9470      7556
+          AV     0.8881    0.7889    0.8356      6722
+          FX     0.9955    0.9928    0.9941      6918
+          NU     0.9684    0.9559    0.9621      6256
+          AJ     0.8974    0.8506    0.8734      4403
+          CL     0.8781    0.8422    0.8598      3739
+          PR     0.8479    0.8523    0.8501      2139
+          NG     1.0000    0.9953    0.9976      1694
+          PA     0.8122    0.8918    0.8501       194
+          XX     0.0000    0.0000    0.0000        27
+          IJ     0.0000    0.0000    0.0000         4
+
+    accuracy                         0.9628    207278
+   macro avg     0.8160    0.8104    0.8128    207278
+weighted avg     0.9624    0.9628    0.9624    207278
+
+lst20 ner_tags best 0.5 0.0
+              precision    recall  f1-score   support
+
+        _BRN     0.4286    0.1915    0.2647        47
+        _DES     0.9090    0.8665    0.8872      1176
+        _DTM     0.7128    0.6657    0.6884      1331
+        _LOC     0.7340    0.6509    0.6900      2349
+        _MEA     0.6669    0.6639    0.6654      3166
+        _NUM     0.6745    0.6267    0.6497      1243
+        _ORG     0.7772    0.6682    0.7186      4261
+        _PER     0.9007    0.8680    0.8840      3272
+        _TRM     0.8835    0.7109    0.7879       128
+        _TTL     0.9673    0.9862    0.9767      1379
+
+   micro avg     0.7873    0.7335    0.7594     18352
+   macro avg     0.7654    0.6898    0.7213     18352
+weighted avg     0.7856    0.7335    0.7579     18352
